added attendance management job

+++ src/sentry/guideSentry.ts
@@ -2,9 +2,9 @@ import { Message, RichEmbed } from 'discord.js';
 
 export const sentinelMessages = (message: Message) => {
   const { content, author } = message;
-  if (!content.includes('http')) {
+  if (!content.includes('http') && message.attachments.size === 0) {
     message.delete();
-    author.send('Please do not chat in the guides channel! If you want to chat about it, please go to general or fox-chat. In case you didn\'t want to lose whatever you sent. This is what you said in guides:');
+    author.send('I\'m sorry! I had to delete your post because we\'re trying to keep the guides channel free from clutter! In case you didn\'t want to lose whatever you sent. This is what you said in guides:');
     author.send(new RichEmbed({
       description: content,
     }));
